azure-image-resize/function_app/ImageResizeFunction/requirements.txt
azure-functions
azure-storage-blob
Pillow
